✨ Feat: serviceAttrSchema
